refactor: extract template editor logic into hook and subcomponent
